Move IListTests
