+++ Tests/Batch4/Batch4.csproj
@@ -101,7 +101,6 @@
     <Compile Include="Reflection\GetMembersTests.cs" />
     <Compile Include="Reflection\ReflectionTests.cs" />
     <Compile Include="Runtime\CompilerServices\RuntimeHelpersTests.cs" />
-    <Compile Include="Collections\Generic\IListTests.cs" />
     <Compile Include="LazyTests.cs" />
     <Compile Include="SimpleTypes\ObjectTests.cs" />
     <Compile Include="Threading\Tasks\PromiseTests.cs" />
